Remove the stripping of HTML for now. Move our replace block with placeholder function into a shared utility
